增加下单相关代码

+++ dao/src/main/java/com/zzlbe/dao/entity/OrderEntity.java
@@ -20,7 +20,7 @@ public class OrderEntity {
     /**
      * 订单编号
      */
-    private Integer orId;
+    private Long orId;
     /**
      * 农化产品id
      */
